Revert back to default views.py
