+++ agileHR/views.py
@@ -2,7 +2,7 @@ import datetime
 from django.shortcuts import get_object_or_404, render
 from django.http import HttpResponse
 
-from agileHR.models import *
+from .models import *
 
 def index(request):
     context = {}
@@ -72,3 +72,38 @@ def traindetail(request, training_id):
     training = get_object_or_404(Training, pk=training_id)
     context = {'training': training}
     return render(request, 'agileHR/training_detail.html', context)
+
+
+def computer(request):
+    """Displays the list of computers currently owned by the company with links to details for each one.
+
+    Author: Sebastian Civarolo
+
+    Returns:
+        render -- loads the computer.html template.
+    """
+
+    computers = Computer.objects.all()
+    context = {
+        "computers": computers
+    }
+    return render(request, 'agileHR/computer.html', context)
+
+
+def computer_detail(request, computer_id):
+    """Displays the details about a single computer owned by the company.
+
+    Arguments:
+        computer_id {int} -- The pk of the computer being requested.
+
+    Returns:
+        render -- Returns the computer_detail template
+    """
+
+    computer = get_object_or_404(Computer, pk=computer_id)
+
+    context = {
+        "computer": computer
+    }
+
+    return render(request, "agileHR/computer_detail.html", context)
