--- agileHR/views/__init__.py
@@ -1,2 +1,0 @@
-from views import *
-from computer_views import *
